Fix previous frame leaking into current frame
Part of https://github.com/marionebl/svg-term-cli/issues/78

All of the frames in a long horizontal strip. Because the frame
text slightly overflows each frame, the content from the last
frame was slightly visible in the current frame.

Essentially, we're adding `overflow: hidden` but in SVG form
so regardless of content in frames, they can't affect each
other.